Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.40.0
python-dotenv==1.0.0
gunicorn==21.2.0
stripe==7.0.0